backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
temporalio
pytest
httpx
